+++ packages/flutter/lib/wenyousite_foundation.dart
@@ -1,4 +1,4 @@
 library;
 
 export 'src/foundation_tokens.dart';
-
+export 'src/wenyou_icons.dart';